Create Clickjacking_detection.html
This code demonstrates clickjacking vulnerability by embedding the target website in an <iframe> tag and setting the width and height of the <iframe> tag to the same width and height as the viewport. This means that the embedded web page will fill the entire viewport, and the user will not be able to see that the website is embedded in an <iframe> tag. As a result, if the user clicks anywhere on the page, they will be clicking on the embedded web page, even though they think they are clicking on the current page.